Open database in read only mode for influx_inspect
influx_inspect would open shards and also enable compactions when it
opened the shards. Since performing compactions should not be something
the influx_inspect tool does, this adds functionality to open an engine
in a read-only mode.

In the future, we can implement more of the methods to follow the mode
that the engine is opened with. Right now, it only controls whether
compactions are enabled or not.

This also disables the internal logging from the library for
influx_inspect. It was obviously intended for this to be the case
because influx_inspect tries to override the logger, but it needed to
use `SetLogOutput` to also override subloggers in the shards.